src/test/java/com/codingquestions/app/RecursionIandSortingAlgorithms/AToThePowerOfBTest.java
package com.codingquestions.app.RecursionIandSortingAlgorithms;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertTrue;

import org.junit.jupiter.api.AfterAll;
import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.BeforeAll;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Disabled;
import org.junit.jupiter.api.Tag;
import org.junit.jupiter.api.Test;

// import the class that you want to test here
import com.codingquestions.app.RecursionIandSortingAlgorithms.AToThePowerOfB;

public class AToThePowerOfBTest {

    @BeforeAll
    static void setup() {
    }

    @BeforeEach
    void setupThis() {
    }

    @Test
    void testMethodOne() {
        AToThePowerOfB instance = new AToThePowerOfB();
        assertEquals(8, instance.methodOne(2, 3));
    }

    @Test
    void testMethodTwo() {
        AToThePowerOfB instance = new AToThePowerOfB();
        assertEquals(8, instance.methodOne(2, 3));
    }

    @AfterEach
    void tearThis() {
    }

    @AfterAll
    static void tear() {
    }
}